Rename db.js to database.js, update imports
